[ALPS06702793] ffa: fix build error
In order to support FF-A, so need to fix below error:

mali_kbase_mem_linux.c:1267:7: error: incompatible pointer types passing u32 * (aka unsigned int *) to parameter of type u64 * (aka unsigned long long *) [-Werror,-Wincompatible-pointer-types]
                             &sec_handle, NULL, 0, 0, &phy_addr);
                             ^~~~~~~~~~~

MTK-Commit-Id: 867b2f3e45c4e6073d73afbec6b0fd685384c2d1

CR-Id: ALPS06702793
Feature: devmpu

+++ gpu_mali/mali_valhall/mali-r38p1/drivers/gpu/arm/midgard/mali_kbase_mem_linux.c
@@ -1254,7 +1254,7 @@ static int kbase_mem_umm_map_attachment(struct kbase_context *kctx,
 
 #if IS_ENABLED(CONFIG_MTK_TRUSTED_MEMORY_SUBSYSTEM) && IS_ENABLED(CONFIG_MTK_GZ_KREE)
 		uint64_t phy_addr = 0;
-		u32 sec_handle = 0;
+		u64 sec_handle = 0;
 
 		if (reg->flags & KBASE_REG_PROTECTED) {
 			struct dma_buf *dma_buf = reg->gpu_alloc->imported.umm.dma_buf;
